Improve structure and documentation of FacetList connector
